Refactor: auth/DTO #39
