+++ src/email/email.controller.ts
@@ -2,6 +2,5 @@ import { Controller } from '@nestjs/common';
 // //TODO: 이메일 인증 시 인증번호를 발송 or 이메일 인증시 링크를 통해 들어온 사용자를 가입 시키는 로직
 // 이쪽에서 처리해야할 로직? 이메일 인증 라우터를 생성후
 // 이메일 인증 처리후 된 값 반환
-햣
 @Controller('email')
 export class EmailController {}
