+++ src/user/dto/index.ts
@@ -1,3 +1,3 @@
 export { CreateUserDto } from './create-user.dto';
-export { LoginUserDto } from './login-user.dto';
+export { JwtDecodeDto } from './jwt-decode-dto';
 export { UpdateUserDto } from './update-user.dto';
