+++ src/user/dto/jwt-decode-dto.ts
@@ -1,0 +1,34 @@
+import { ApiProperty } from '@nestjs/swagger';
+import {
+  IsEmail,
+  IsIn,
+  IsNotEmpty,
+  IsOptional,
+  IsString,
+} from 'class-validator';
+
+export class JwtDecodeDto {
+  @ApiProperty()
+  @IsString()
+  @IsEmail()
+  readonly email: string;
+
+  @ApiProperty()
+  @IsString()
+  @IsNotEmpty()
+  readonly password: string;
+  @ApiProperty()
+  readonly accessToken: string;
+
+  @ApiProperty()
+  readonly refreshToken: string;
+}
+
+export interface JwtDecodeDto {
+  [key: string]: string | number;
+  id: number;
+  email: string;
+  nickname: string;
+  iat: number;
+  exp: number;
+}
